fix: instance manager kills the correct instance
refactor: move telescope to an extension folder
          disable telescope

+++ lua/godoterm/actions/close_editor.lua
@@ -18,7 +18,9 @@ local function close_editor()
                 return fmt
             end
         }, function (_,idx)
-            instance_m.kill(instance_m.remove(idx))
+            if idx then
+                instance_m.kill(instance_m.remove(idx))
+            end
         end)
     end
